added license of original code
